requirements.txt
asgiref==3.7.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
Django==4.2.5